Proper error codes and messages added

+++ src/main/java/bloodbank/rest/resource/AddressResource.java
@@ -18,6 +18,7 @@ import javax.ws.rs.core.Response.Status;
 import javax.annotation.security.RolesAllowed;
 import javax.ejb.EJB;
 import javax.inject.Inject;
+import javax.persistence.NoResultException;
 import javax.security.enterprise.SecurityContext;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -55,7 +56,7 @@ public class AddressResource {
 	public Response getAddresses() {
 		LOG.debug( "retrieving all addresses ...");
 		List< Address> addresses = service.getAllAddresses();
-		Response response = Response.ok( addresses).build();
+		Response response = Response.status( addresses == null || addresses.isEmpty() ? Status.NOT_FOUND : Status.OK).entity( addresses).build();
 		return response;
 	}
 
@@ -65,6 +66,10 @@ public class AddressResource {
 	public Response getAddressById( @PathParam( RESOURCE_PATH_ID_ELEMENT) int id) {
 		LOG.debug( "try to retrieve specific address " + id);	
 		Address address = service.getAddressById(id);
+		if(address == null) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.NOT_FOUND.getStatusCode(), "No Address found with id " + id);
+			return Response.status(Status.NOT_FOUND).entity(error).build();
+		}
 		Response response = Response.status( address == null ? Status.NOT_FOUND : Status.OK).entity( address).build();
 		return response;
 	}
@@ -73,8 +78,44 @@ public class AddressResource {
 	@RolesAllowed( { ADMIN_ROLE })
 	public Response addAddress( Address newAddress) {
 		LOG.debug( "add a new address");
+		Response response = null;
+		if(newAddress.getCountry() == null || newAddress.getCountry().equals("")) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.BAD_REQUEST.getStatusCode(), "Country is missing");
+			return Response.status(Status.BAD_REQUEST).entity(error).build();
+		}
+		
+		if(newAddress.getProvince() == null || newAddress.getProvince().equals("")) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.BAD_REQUEST.getStatusCode(), "Province is missing");
+			return Response.status(Status.BAD_REQUEST).entity(error).build();
+		}
+		
+		if(newAddress.getCity() == null || newAddress.getCity().equals("")) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.BAD_REQUEST.getStatusCode(), "City is missing");
+			return Response.status(Status.BAD_REQUEST).entity(error).build();
+		}
+		
+		if(newAddress.getStreet() == null || newAddress.getStreet().equals("")) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.BAD_REQUEST.getStatusCode(), "Street is missing");
+			return Response.status(Status.BAD_REQUEST).entity(error).build();
+		}
+		
+		if(newAddress.getStreet() == null || newAddress.getStreet().equals("")) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.BAD_REQUEST.getStatusCode(), "Street is missing");
+			return Response.status(Status.BAD_REQUEST).entity(error).build();
+		} 
+		
+		if(newAddress.getStreetNumber() == null || newAddress.getStreetNumber().equals("")) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.BAD_REQUEST.getStatusCode(), "Street Number is missing");
+			return Response.status(Status.BAD_REQUEST).entity(error).build();
+		} 
+		
+		if(newAddress.getZipcode() == null || newAddress.getZipcode().equals("")) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.BAD_REQUEST.getStatusCode(), "ZIP Code is missing");
+			return Response.status(Status.BAD_REQUEST).entity(error).build();
+		} 
+			
 		Address addedAddress = service.persistAddress( newAddress);
-		Response response = Response.ok( addedAddress).build();
+		response = Response.ok( addedAddress).build();
 		return response;
 	}	
 	
@@ -82,6 +123,11 @@ public class AddressResource {
 	@RolesAllowed( { ADMIN_ROLE })
 	@Path( RESOURCE_PATH_ID_PATH)
 	public Response deleteAddressById( @PathParam( RESOURCE_PATH_ID_ELEMENT) int id) {
+		Address address = service.getAddressById(id);
+		if(address == null) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.NOT_FOUND.getStatusCode(), "No Address found with id " + id);
+			return Response.status(Status.NOT_FOUND).entity(error).build();
+		}
 		Address deletedAddress = service.deleteAddressById(id);
 		Response response = Response.ok( deletedAddress).build();
 		return response;
@@ -93,6 +139,10 @@ public class AddressResource {
 	public Response updateAddressById( @PathParam( RESOURCE_PATH_ID_ELEMENT) int id,  Address updatingAddress) {
 		Response response = null;
 		Address updatedAddress = service.updateAddressById(id, updatingAddress);
+		if(updatedAddress == null) {
+			HttpErrorResponse error = new HttpErrorResponse(Status.NOT_FOUND.getStatusCode(), "No Address found with id " + id);
+			return Response.status(Status.NOT_FOUND).entity(error).build();
+		}
 		response = Response.ok( updatedAddress).build();
 		return response;
 	}
